streamlit_dashboard/requirements.txt
streamlit
paho-mqtt
pymongo
pandas

